% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-curve-outline.R
\name{geom_curve_outline}
\alias{geom_curve_outline}
\title{Hilbert curve outline}
\usage{
geom_curve_outline(
  mapping = NULL,
  data = NULL,
  ...,
  enclosed = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{enclosed}{A logical scalar indicating whether the curve outline should
have closed ends (default: \code{FALSE}).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
Hilbert curve outline
}
\section{Aesthetics}{

\code{geom_curve_outline()} understands the following aesthetics:
\itemize{
\item \code{ip}
\item \code{curve_order}
\item \code{alpha}
\item \code{colour}
\item \code{linetype}
\item \code{size}
}
}

