% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-ip.R
\name{coord_ip}
\alias{coord_ip}
\title{Map IP data onto Cartesian coordinate system}
\usage{
coord_ip(
  network = ip_network("0.0.0.0/0"),
  pixel_prefix = 16L,
  curve = c("hilbert", "morton")
)
}
\arguments{
\item{network}{An \code{\link[ipaddress:ip_network]{ipaddress::ip_network}} scalar}

\item{pixel_prefix}{An integer scalar}

\item{curve}{A string indicating the space-filling curve used to map IP data
to the Cartesian plane. Choices are \code{"hilbert"} (default) and \code{"morton"}.}
}
\description{
Map IP data onto Cartesian coordinate system
}
