% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggip-package.R
\docType{package}
\name{ggip-package}
\alias{ggip}
\alias{ggip-package}
\title{ggip: Data Visualization for IP Addresses and Networks}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A 'ggplot2' extension that enables visualization of
    IP (Internet Protocol) addresses and networks. The address space is
    mapped onto the Cartesian coordinate system using a space-filling
    curve. Offers full support for both IPv4 and IPv6 (Internet Protocol
    versions 4 and 6) address spaces.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidchall.github.io/ggip/}
  \item \url{https://github.com/davidchall/ggip}
  \item Report bugs at \url{https://github.com/davidchall/ggip/issues}
}

}
\author{
\strong{Maintainer}: David Hall \email{david.hall.physics@gmail.com} (\href{https://orcid.org/0000-0002-2193-0480}{ORCID})

}
\keyword{internal}
