% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_to_cartesian.R
\name{ip_to_cartesian}
\alias{ip_to_cartesian}
\alias{address_to_cartesian}
\alias{network_to_cartesian}
\title{Map IP data to Cartesian coordinates}
\usage{
address_to_cartesian(
  address,
  canvas_network = ip_network("0.0.0.0/0"),
  pixel_prefix = 16,
  curve = "hilbert"
)

network_to_cartesian(
  network,
  canvas_network = ip_network("0.0.0.0/0"),
  pixel_prefix = 16,
  curve = "hilbert"
)
}
\arguments{
\item{canvas_network}{An \code{\link[ipaddress:ip_network]{ipaddress::ip_network}} scalar that determines the
network visualized by the plotted canvas. The default shows the entire IPv4
address space.}

\item{pixel_prefix}{An integer scalar that determines the number of addresses
represented by a single pixel. It sets the prefix length of the
corresponding network. The default value is 16.}

\item{curve}{A string to choose the space-filling curve. Choices are
\code{"hilbert"} (default) and \code{"morton"}.}
}
\description{
Map IP data to Cartesian coordinates
}
